sore/sore/asdfghjkl;ASDFGHJKL/normalisation_service/requirements.txt
aiokafka>=0.8.0
pydantic>=2.0,<3.0
pyyaml>=6.0
structlog>=24.0.0
prometheus-client>=0.17.0
pytest>=8.0
pytest-asyncio>=0.21.0